fixed a type error in the previous commit

+++ compilers/segmented-memory-backend/src/utils/interval.cc
@@ -452,7 +452,7 @@ MultidimensionalIntervalSeq::~MultidimensionalIntervalSeq() {
         }
 }
 
-int MultidimensionalIntervalSeq::getNumOfElements() {
+long int MultidimensionalIntervalSeq::getNumOfElements() {
 	long int count = intervals[0]->getNumOfElements();
 	for (int i = 1; i < dimensionality; i++) {
 		count *= intervals[i]->getNumOfElements();
